Adds initial support for multiple OpenCL devices
Adds support to MMD for multiple devices and modifies the
"aocl diagnose" and "aocl program" tools to support this
change.

The MMD will return results for BSP that are loaded when
using the existing aocl_mmd calls.  These calls will
only work for devices that have an existing aocx with
BSP loaded as the AFU in the green bits.

Two additional ccip_mmd calls are added that are used by
"aocl diagnose" and "aocl program" to get device status
and load aocx for devices that don't yet have an OpenCL AFU
with BSP loaded . This allows users to convert an AFU to
have a BSP.  The number of active devices reported by
standard aocl calls will match with number of devices
loaded with a BSP.

Current version uses OPAE ObjectID directly for naming devices.
This may not be user friendly enough, but still working to
determine what the best alternative is.