Cleanup and centralize parsing of device name
Device name parsing is centralized in CcipDevice rather
than being split across several functions and two files.  This
should simplify code maintenance when parsing function changes.

The parsing still needs to be improved but that will be done
as a follow-on patch

This patch also updates the MMD info string to print device BDF.