Remove unnecessary and insufficient locking
The aocl_mmd function calls are not thread-safe.  Instead they rely
on the aoc runtime to enforce synchronization.  Initially thought
that locking should be added for creating CcipDevice.  But this
locking is insufficient due to other parts of code that modify
shared data, and so it is removed here.

Also added comment that the functions are not thread-safe.