Return temperature from aocl_mmd_get_info
Read the device temperature from sysfs and return result
when aocl_mmd_get_info requests temperature reading.

This implementation relies on the fact that the objectID
is constructed using the device instance ID as the lower
20 bits of its value.  This implementation also relies
on a hard-coded path to the temperature reading in sysfs.

This should work fine for the current implementation of
OPAE driver and SDK.  However, it probably won't work in
the future.

Once OPAE is able to return the temperature value directly
this should be modified to use an OPAE C API call.  That
doesn't currently exist so reading from sysfs directly
is a workaround for DCP 1.0 timeframe.