Update dcp name and fix diagnose return code
Change the device prefix name from dcp to
pac_a10.

This change will be in place for beta.  May revisit
prior to PRQ to see if we need to specify a different
name for inspur.

This change also improves parsing somewhat.  Should
be good enough for beta, but should also be tested
more throughly before PRQ.

Also fixes klocwork issue in eventfd_wrapper
where return value from write had wrong type.

With this commit all the important issues
reported by klocwork are fixed.  Remaining
ones are "not a problem" though may be
areas to refactor code in the future.