Use memcpy_s by default
Initial integration of memcpy_s in MMD resulted in large performance
hit. As a result memcpy_s was replaced with a macro allowing for it
to be easily enabled/disabled.

The performance hit has since been mitigated through changes in the
DMA code.  This patch re-enables memcpy_s in the remaining places
where it is used.  It retains the macro for enable/disable for now
and simply enables memcpy_s.