Use logical names for aocl diagnose and program
After switching the AOC 17.1 RTE we learned that OpenCL does not
support use of physical device names.  Instead we should use
logical device names.

This commit changes to treat an DCP device that does not have a
BSP as correclty configured device.  This assigns it a logical
name that can be used by 'aocl program'.