Add lib64 as possible OPAE install location
Currently the regtests for OpenCL source the path that is set
by the 'opencl bash' subcommand.  Recently this changed from
$OPAE_INSTALL/lib to $OPAE_INSTALL/lib64.

Long term the regtests should probably not depend on a special
case in opencl-bsp repo.

Short term this is probably the fastest fix.  Adding lib64 in
case there are still other things that depend on the lib location.